black reformat

+++ notebooks/Deep_Learning_listmode_PET/snippets/solution_2_1.py
@@ -1,7 +1,7 @@
 # we first create an numpy array that we fill in a loop
-numpy_image_3: np.ndarray = np.zeros(sirf_image_1.shape, dtype = sirf_image_1.dtype)
+numpy_image_3: np.ndarray = np.zeros(sirf_image_1.shape, dtype=sirf_image_1.dtype)
 for i in range(numpy_image_3.shape[0]):
-    numpy_image_3[i,:,:] = i**2
+    numpy_image_3[i, :, :] = i ** 2
 
 sirf_image_3: sirf.STIR.ImageData = acq_data.create_uniform_image(0.0)
 sirf_image_3.fill(numpy_image_3)
